wsd: fix support-key build
